GameManager added
The drafting mechanism is added. After drafting, the player can decide to build or recycle the cards. The recycled materials are displayed with a UI element, that isn't visible in the drafting period. It is possible to build cards if there is enough material from recycling, however individual material putting is not possible yet.